将strToInt的返回值修改为long long，以满足32位系统下更长整形的要求

+++ tools/stringtool.cpp
@@ -137,7 +137,7 @@ void replaceAllSubStrM(std::string &source,
 }
 
 
-long strToInt(const std::string &source) {
+long long strToInt(const std::string &source) {
   if(!isIntegerNumber(source)) {
     return long(0);
   } else {
